minor refactor of agents and isorting
