+++ src/crl/cons/_type.py
@@ -1,0 +1,5 @@
+from typing import Literal
+
+ClassicControl = Literal[
+    "CartPole-v1", "Acrobot-v1", "LunarLander-v3", "MountainCar-v0"
+]
