+++ src/crl/cons/agents/__init__.py
@@ -1,0 +1,3 @@
+from crl.cons.agents.cql import learn_cqldqn_policy
+from crl.cons.agents.ddqn import learn_ddqn_policy
+from crl.cons.agents.dqn import learn_dqn_policy
